:art: refactor: reorganize handlers
