+++ api/app/routers/land_change/tree_cover_gain/tree_cover_gain.py
@@ -3,7 +3,11 @@ from app.domain.analyzers.dummy_tree_cover_gain_analyzer import (
 )
 from app.domain.compute_engines.compute_engine import (
     ComputeEngine,
+)
+from app.domain.compute_engines.handlers.otf_implementations.flox_otf_handler import (
     FloxOTFHandler,
+)
+from app.domain.compute_engines.handlers.precalc_implementations.precalc_handlers import (
     TreeCoverGainPrecalcHandler,
 )
 from app.domain.compute_engines.handlers.precalc_implementations.precalc_sql_query_builder import (
