+++ api/app/routers/land_change/tree_cover_loss/tree_cover_loss.py
@@ -1,7 +1,11 @@
 from app.domain.analyzers.tree_cover_loss_analyzer import TreeCoverLossAnalyzer
 from app.domain.compute_engines.compute_engine import (
     ComputeEngine,
+)
+from app.domain.compute_engines.handlers.otf_implementations.flox_otf_handler import (
     FloxOTFHandler,
+)
+from app.domain.compute_engines.handlers.precalc_implementations.precalc_handlers import (
     TreeCoverLossPrecalcHandler,
 )
 from app.domain.compute_engines.handlers.precalc_implementations.precalc_sql_query_builder import (
