+++ api/test/integration/domain/compute_engines/test_precalc_handler.py
@@ -1,7 +1,7 @@
 from unittest.mock import MagicMock
 
 import pytest
-from app.domain.compute_engines.compute_engine import (
+from app.domain.compute_engines.handlers.precalc_implementations.precalc_handlers import (
     TreeCoverGainPrecalcHandler,
 )
 from app.domain.compute_engines.handlers.precalc_implementations.precalc_sql_query_builder import (
